// packages/core/src/db/database-generated.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  wallet: {
    Tables: {
      accounts: {
        Row: {
          created_at: string;
          currency: Database['wallet']['Enums']['currency'];
          details: Json;
          id: string;
          name: string;
          purpose: Database['wallet']['Enums']['account_purpose'];
          type: Database['wallet']['Enums']['account_type'];
          user_id: string;
          version: number;
        };
        Insert: {
          created_at?: string;
          currency: Database['wallet']['Enums']['currency'];
          details: Json;
          id?: string;
          name: string;
          purpose?: Database['wallet']['Enums']['account_purpose'];
          type: Database['wallet']['Enums']['account_type'];
          user_id: string;
          version?: number;
        };
        Update: {
          created_at?: string;
          currency?: Database['wallet']['Enums']['currency'];
          details?: Json;
          id?: string;
          name?: string;
          purpose?: Database['wallet']['Enums']['account_purpose'];
          type?: Database['wallet']['Enums']['account_type'];
          user_id?: string;
          version?: number;
        };
        Relationships: [
          {
            foreignKeyName: 'accounts_user_id_fkey';
            columns: ['user_id'];
            isOneToOne: false;
            referencedRelation: 'users';
            referencedColumns: ['id'];
          },
        ];
      };
      cashu_proofs: {
        Row: {
          account_id: string;
          amount: string;
          cashu_receive_quote_id: string | null;
          cashu_receive_swap_token_hash: string | null;
          cashu_send_quote_id: string | null;
          cashu_send_swap_id: string | null;
          created_at: string;
          dleq: Json | null;
          id: string;
          keyset_id: string;
          public_key_y: string;
          reserved_at: string | null;
          secret: string;
          spending_cashu_send_quote_id: string | null;
          spending_cashu_send_swap_id: string | null;
          spent_at: string | null;
          state: Database['wallet']['Enums']['cashu_proof_state'];
          unblinded_signature: string;
          user_id: string;
          version: number;
          witness: Json | null;
        };
        Insert: {
          account_id: string;
          amount: string;
          cashu_receive_quote_id?: string | null;
          cashu_receive_swap_token_hash?: string | null;
          cashu_send_quote_id?: string | null;
          cashu_send_swap_id?: string | null;
          created_at?: string;
          dleq?: Json | null;
          id?: string;
          keyset_id: string;
          public_key_y: string;
          reserved_at?: string | null;
          secret: string;
          spending_cashu_send_quote_id?: string | null;
          spending_cashu_send_swap_id?: string | null;
          spent_at?: string | null;
          state?: Database['wallet']['Enums']['cashu_proof_state'];
          unblinded_signature: string;
          user_id: string;
          version?: number;
          witness?: Json | null;
        };
        Update: {
          account_id?: string;
          amount?: string;
          cashu_receive_quote_id?: string | null;
          cashu_receive_swap_token_hash?: string | null;
          cashu_send_quote_id?: string | null;
          cashu_send_swap_id?: string | null;
          created_at?: string;
          dleq?: Json | null;
          id?: string;
          keyset_id?: string;
          public_key_y?: string;
          reserved_at?: string | null;
          secret?: string;
          spending_cashu_send_quote_id?: string | null;
          spending_cashu_send_swap_id?: string | null;
          spent_at?: string | null;
          state?: Database['wallet']['Enums']['cashu_proof_state'];
          unblinded_signature?: string;
          user_id?: string;
          version?: number;
          witness?: Json | null;
        };
        Relationships: [
          {
            foreignKeyName: 'cashu_proofs_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'cashu_proofs_cashu_receive_quote_id_fkey';
            columns: ['cashu_receive_quote_id'];
            isOneToOne: false;
            referencedRelation: 'cashu_receive_quotes';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'cashu_proofs_cashu_send_quote_id_fkey';
            columns: ['cashu_send_quote_id'];
            isOneToOne: false;
            referencedRelation: 'cashu_send_quotes';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'cashu_proofs_cashu_send_swap_id_fkey';
            columns: ['cashu_send_swap_id'];
            isOneToOne: false;
            referencedRelation: 'cashu_send_swaps';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'cashu_proofs_receive_swap_fkey';
            columns: ['cashu_receive_swap_token_hash', 'user_id'];
            isOneToOne: false;
            referencedRelation: 'cashu_receive_swaps';
            referencedColumns: ['token_hash', 'user_id'];
          },
          {
            foreignKeyName: 'cashu_proofs_spending_cashu_send_quote_id_fkey';
            columns: ['spending_cashu_send_quote_id'];
            isOneToOne: false;
            referencedRelation: 'cashu_send_quotes';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'cashu_proofs_spending_cashu_send_swap_id_fkey';
            columns: ['spending_cashu_send_swap_id'];
            isOneToOne: false;
            referencedRelation: 'cashu_send_swaps';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'cashu_proofs_user_id_fkey';
            columns: ['user_id'];
            isOneToOne: false;
            referencedRelation: 'users';
            referencedColumns: ['id'];
          },
        ];
      };
      cashu_receive_quotes: {
        Row: {
          account_id: string;
          cashu_token_melt_initiated: boolean | null;
          created_at: string;
          encrypted_data: string;
          expires_at: string;
          failure_reason: string | null;
          id: string;
          keyset_counter: number | null;
          keyset_id: string | null;
          locking_derivation_path: string;
          payment_hash: string;
          quote_id_hash: string;
          state: Database['wallet']['Enums']['cashu_receive_quote_state'];
          transaction_id: string;
          type: Database['wallet']['Enums']['receive_quote_type'];
          user_id: string;
          version: number;
        };
        Insert: {
          account_id: string;
          cashu_token_melt_initiated?: boolean | null;
          created_at?: string;
          encrypted_data: string;
          expires_at: string;
          failure_reason?: string | null;
          id?: string;
          keyset_counter?: number | null;
          keyset_id?: string | null;
          locking_derivation_path: string;
          payment_hash: string;
          quote_id_hash: string;
          state: Database['wallet']['Enums']['cashu_receive_quote_state'];
          transaction_id: string;
          type: Database['wallet']['Enums']['receive_quote_type'];
          user_id: string;
          version?: number;
        };
        Update: {
          account_id?: string;
          cashu_token_melt_initiated?: boolean | null;
          created_at?: string;
          encrypted_data?: string;
          expires_at?: string;
          failure_reason?: string | null;
          id?: string;
          keyset_counter?: number | null;
          keyset_id?: string | null;
          locking_derivation_path?: string;
          payment_hash?: string;
          quote_id_hash?: string;
          state?: Database['wallet']['Enums']['cashu_receive_quote_state'];
          transaction_id?: string;
          type?: Database['wallet']['Enums']['receive_quote_type'];
          user_id?: string;
          version?: number;
        };
        Relationships: [
          {
            foreignKeyName: 'cashu_receive_quotes_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'cashu_receive_quotes_transaction_id_fkey';
            columns: ['transaction_id'];
            isOneToOne: false;
            referencedRelation: 'transactions';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'cashu_receive_quotes_user_id_fkey';
            columns: ['user_id'];
            isOneToOne: false;
            referencedRelation: 'users';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
      users: {
        Row: {
          cashu_locking_xpub: string;
          created_at: string;
          default_btc_account_id: string | null;
          default_currency: Database['wallet']['Enums']['currency'];
          default_usd_account_id: string | null;
          email: string | null;
          email_verified: boolean;
          encryption_public_key: string;
          id: string;
          spark_identity_public_key: string;
          terms_accepted_at: string | null;
          updated_at: string;
          username: string;
        };
        Insert: {
          cashu_locking_xpub: string;
          created_at?: string;
          default_btc_account_id?: string | null;
          default_currency?: Database['wallet']['Enums']['currency'];
          default_usd_account_id?: string | null;
          email?: string | null;
          email_verified: boolean;
          encryption_public_key: string;
          id?: string;
          spark_identity_public_key: string;
          terms_accepted_at?: string | null;
          updated_at?: string;
          username: string;
        };
        Update: {
          cashu_locking_xpub?: string;
          created_at?: string;
          default_btc_account_id?: string | null;
          default_currency?: Database['wallet']['Enums']['currency'];
          default_usd_account_id?: string | null;
          email?: string | null;
          email_verified?: boolean;
          encryption_public_key?: string;
          id?: string;
          spark_identity_public_key?: string;
          terms_accepted_at?: string | null;
          updated_at?: string;
          username?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'users_default_btc_account_id_fkey';
            columns: ['default_btc_account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'users_default_usd_account_id_fkey';
            columns: ['default_usd_account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
        ];
      };
    };
// ...
  };
};
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema['Enums']
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
    ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
    : never;